+++ projects/ngx-material/src/directives/index.ts
@@ -1,3 +1,4 @@
 
 export * from './context-menu.directive';
-export * from './wrap-content.directive';
+export * from './ui-wrapper.directive';
+export * from './ui-header.directive';