+++ projects/ngx-tools/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @juulsgaard/ngx-tools
 
+## 16.1.29
+
+### Patch Changes
+
+- Stop time based directives when in background
+
+## 16.1.28
+
+### Patch Changes
+
+- Update drag directives to handle Safari weirdness
+
 ## 16.1.27
 
 ### Patch Changes